Fix poor styling of textareas inside the Django admin app.
The Django admin app's default stylesheet includes a blanket rule that
limits the height of all textarea elements, if the browser window is
less than 1024 pixels wide.

This interferes with TinyMCE's use of a textarea element in the
"Source Code" dialog.  In that dialog, the textarea (.tox-textarea) is
placed inside a div wrapper (.tox-textarea-wrap); due to TinyMCE's
stylesheet, the div has a visible border while the textarea itself has
none.  So the overall effect is that if you click the "Source Code"
button, a large dialog box appears (covering most of the browser
window), but only the top 120 pixels of that box are usable.  This is
visually confusing and makes poor use of screen space.

Fix this by adding a small additional stylesheet to the AdminTinyMCE
widget.